feat: 3-Agent 제거와 LLM 경계 재배치
파싱 경계를 이중화했다. parsing-preview가 규칙 파서와 LLM 파서를 나란히 돌리고
불일치 필드를 응답에 실어 사람이 고르게 한다. 어느 쪽도 코드가 자동 채택하지
않는다 — 그 자리에서 코드가 고른 값은 다시 검증되지 않기 때문이다. 키가 없으면
규칙 파서 결과만 담고 그 사실을 표시한다. 저하는 되지만 실패하지는 않는다.

안내문을 LLM 한 번 호출로 줄였다. 판정·DSR·추천 상세가 이미 DB에 있는데
파이프라인이 같은 심사를 다시 하고 그 결과를 버리고 있었다. 확정값을 프롬프트에
데이터로 넣으면 도구 왕복 네 번이 사라진다. 보장도 함께 올라간다 — "도구를
부르라고 프롬프트로 시키는 것"과 "생성할 대상이 애초에 없는 것"은 다른 층위다.

안내문을 실행하는 주체를 키의 출처로 갈랐다. 서버 키는 워커가 비동기로, 방문자
키는 app이 동기로 쓴다. 방문자 키가 HTTP 요청으로 닿을 수 있는 가장 안쪽이
거기이고, 더 안으로 보내려면 저장해야 하는데 키 저장은 금지다. 두 경로가 같은
행을 집고 같은 컬럼을 채우므로 감사 경로는 하나로 남는다. 심사 생성이 PENDING을
함께 만들므로 동기 경로는 그 행을 집어야 한다 — 409로 막으면 방문자가 항상
409를 받는다. 운영에 워커가 없어 스테일 RUNNING을 회수할 주체도 없으므로 그
회수까지 동기 경로가 맡는다.

3-Agent 코드와 노트북을 지웠다. 경계 세 개 중 둘이 재구축으로 비었고 남은 코드는
어디서도 불리지 않았다. 노트북은 목차와 절 제목뿐이라 살릴 산문이 없었고 뼈대가
CrewAI 3-Agent라 고쳐 쓸 대상도 아니었다. 이전 설계는 git이 보존한다.

스타일·LLM 경로·계약 상수를 파일 경계로 분리했다. core.py에 LLM을 부르는 코드가
한 줄도 남지 않았다 — "판정은 코드, 설명은 LLM"이 주석이 아니라 디렉터리 구조로
확인된다. CSS 217줄은 화면 로직과 다른 이유로 바뀌므로 파일로 뺐다.

코드 리뷰 지적 일곱 건을 고쳤다. 그중 하나는 워커를 죽이는 결함이었다 — 타임아웃
분기에 return이 없어 할당된 적 없는 변수로 흘러내렸고 그 예외가 실행 루프까지
올라갔다. 토큰이 항상 None이던 원인도 여기서 나왔다. 삼항 조건이 전체 식을 dict인
경우로 묶어, 객체를 처리하려던 분기가 도달할 수 없었다. 둘 다 회귀 테스트를 붙였고
고친 것을 되돌리면 그 테스트가 깨지는지 확인했다.

dsr 저장 정밀도를 명시했다. Numeric은 읽을 때 Decimal을 돌려주는데 타입 힌트는
float였다. 왕복 타입을 단언하는 통합 테스트를 함께 넣었다. 판정 계산은 여전히
float다 — 밴드 경계에서 실측해 판정이 갈리지 않음을 확인했다.

화면 라벨을 실제 구조로 바꿨다. 삭제된 Agent 1/2/3을 계속 표시하고 있었고, 특히
"Agent 2 — 심사 결과"는 심사를 LLM이 한다고 읽혀 이 서비스의 주장과 정면으로
어긋났다.

+++ loan_agent/db/models.py
@@ -9,6 +9,7 @@
 """
 import datetime
 import uuid
+from decimal import Decimal
 
 from sqlalchemy import (
     BigInteger,
@@ -104,7 +105,9 @@ class DecisionResult(Base):
     )
     verdict: Mapped[str] = mapped_column(Text, nullable=False)
     repayment_band: Mapped[str] = mapped_column(Text, nullable=False)
-    dsr: Mapped[float] = mapped_column(Numeric, nullable=False)
+    # Numeric은 DB 왕복 시 Decimal을 돌려준다. 판정 계산은 float로 유지하되 저장 경계의
+    # 타입을 숨기지 않고, DSR 비교에 충분한 소수 12자리 정밀도를 명시한다(ADR-032).
+    dsr: Mapped[Decimal] = mapped_column(Numeric(18, 12), nullable=False)
     monthly_payment: Mapped[dict] = mapped_column(JSONB, nullable=False)
     rule_version: Mapped[str] = mapped_column(Text, nullable=False)
     product_dataset_version: Mapped[str] = mapped_column(Text, nullable=False)
